Create Array .c
This C program allows the user to enter a list of integers and then counts how many of them are positive, negative, or zero. It's a great beginner-level exercise that demonstrates how to:

Use arrays for storing user input

Traverse arrays with loops

Use conditional statements (if, else if, else)

Maintain counters based on conditions


Use Case:
This program is helpful when analyzing datasets or user input where categorization of numbers is needed.