Create Average of array elements .c
This C program calculates the average (mean) of all elements in an array. It demonstrates the use of arrays, loops, and basic arithmetic operations. This is a great starting point for understanding how data can be processed for statistical insights.


---

Key Concepts:

Looping through arrays

Accumulating a total

Computing average using division